hybrid works
