+++ advanced-analytics-hybrid/generate_data/create_base_table.sql
