+++ advanced-analytics-hybrid/generate_data/create_final_table.sql
