+++ advanced-analytics-hybrid/streamlit/data/avg_duration.csv
@@ -1,0 +1,2 @@
+average_session_session_duration_s
+8
