+++ advanced-analytics-hybrid/streamlit/data/avg_duration_by_day.csv
@@ -1,0 +1,151 @@
+date,avg_session_duration
+2021-09-21,5
+2021-09-25,11
+2021-09-28,11
+2021-09-29,11
+2021-10-02,11
+2021-10-05,10
+2021-10-16,11
+2021-10-17,0
+2021-10-18,10
+2021-10-20,7
+2021-10-22,11
+2021-10-23,11
+2021-10-27,5
+2021-11-01,6
+2021-11-08,11
+2021-11-12,11
+2021-11-13,11
+2021-11-17,11
+2021-11-20,10
+2021-11-25,11
+2021-11-28,5
+2021-12-05,10
+2021-12-15,5
+2021-12-24,6
+2021-12-25,11
+2021-12-26,10
+2021-12-28,5
+2022-01-05,12
+2022-01-07,10
+2022-01-10,0
+2022-01-15,0
+2022-01-28,5
+2022-01-30,11
+2022-02-02,5
+2022-02-13,10
+2022-02-19,12
+2022-02-20,10
+2022-02-22,11
+2022-02-24,11
+2022-03-12,10
+2022-03-19,11
+2022-03-21,11
+2022-03-25,11
+2022-03-26,11
+2022-03-29,2
+2022-04-01,10
+2022-04-03,7
+2022-04-04,3
+2022-04-05,2
+2022-04-06,8
+2022-04-09,5
+2022-04-12,4
+2022-04-14,10
+2022-04-15,12
+2022-04-17,5
+2022-04-18,10
+2022-04-19,10
+2022-04-21,10
+2022-04-22,10
+2022-04-26,5
+2022-05-01,10
+2022-05-03,11
+2022-05-05,11
+2022-05-06,7
+2022-05-07,12
+2022-05-08,0
+2022-05-10,11
+2022-05-11,11
+2022-05-12,5
+2022-05-14,0
+2022-05-15,5
+2022-05-16,10
+2022-05-18,11
+2022-05-19,12
+2022-05-20,10
+2022-05-22,11
+2022-05-23,10
+2022-05-24,11
+2022-05-25,6
+2022-05-26,12
+2022-05-28,11
+2022-05-31,11
+2022-06-01,11
+2022-06-03,2
+2022-06-04,11
+2022-06-05,10
+2022-06-06,6
+2022-06-07,12
+2022-06-08,0
+2022-06-09,11
+2022-06-10,11
+2022-06-11,11
+2022-06-13,10
+2022-06-15,12
+2022-06-17,2
+2022-06-20,10
+2022-06-22,11
+2022-06-24,12
+2022-06-25,5
+2022-06-26,12
+2022-06-27,1
+2022-07-02,8
+2022-07-04,11
+2022-07-05,5
+2022-07-06,0
+2022-07-08,11
+2022-07-09,8
+2022-07-10,6
+2022-07-13,1
+2022-07-14,0
+2022-07-15,0
+2022-07-17,10
+2022-07-18,6
+2022-07-19,11
+2022-07-20,7
+2022-07-21,10
+2022-07-22,11
+2022-07-23,10
+2022-07-24,11
+2022-07-25,0
+2022-07-27,8
+2022-07-29,11
+2022-07-31,11
+2022-08-01,11
+2022-08-02,10
+2022-08-04,12
+2022-08-05,2
+2022-08-06,11
+2022-08-09,10
+2022-08-10,10
+2022-08-12,5
+2022-08-13,11
+2022-08-14,10
+2022-08-16,6
+2022-08-19,7
+2022-08-20,10
+2022-08-21,12
+2022-08-23,11
+2022-08-24,0
+2022-08-25,10
+2022-08-27,12
+2022-08-28,10
+2022-08-29,11
+2022-08-31,5
+2022-09-02,0
+2022-09-03,12
+2022-09-04,10
+2022-09-05,12
+2022-09-09,0
+2022-09-11,10
