+++ advanced-analytics-hybrid/streamlit/data/bounce_rate_by_day.csv
@@ -1,0 +1,33 @@
+date,bouncerate
+2021-10-17,1.000000
+2021-11-01,0.250000
+2021-12-24,0.142857
+2022-01-10,1.000000
+2022-01-15,1.000000
+2022-01-28,0.142857
+2022-04-03,0.100000
+2022-04-04,0.400000
+2022-04-12,0.100000
+2022-05-06,0.111111
+2022-05-08,1.000000
+2022-05-14,1.000000
+2022-05-25,0.100000
+2022-06-06,0.142857
+2022-06-08,1.000000
+2022-06-25,0.142857
+2022-06-27,0.333333
+2022-07-06,1.000000
+2022-07-09,0.058824
+2022-07-10,0.250000
+2022-07-13,0.333333
+2022-07-14,1.000000
+2022-07-15,1.000000
+2022-07-18,0.100000
+2022-07-20,0.076923
+2022-07-25,1.000000
+2022-07-27,0.066667
+2022-08-12,0.142857
+2022-08-16,0.166667
+2022-08-24,1.000000
+2022-09-02,1.000000
+2022-09-09,1.000000
