+++ advanced-analytics-hybrid/streamlit/data/num_sessions.csv
@@ -1,0 +1,2 @@
+number_of_sessions
+210
