+++ advanced-analytics-hybrid/streamlit/data/num_sessions_by_day.csv
@@ -1,0 +1,151 @@
+date,number_of_sessions
+2021-09-21,1
+2021-09-25,1
+2021-09-28,1
+2021-09-29,1
+2021-10-02,1
+2021-10-05,1
+2021-10-16,1
+2021-10-17,1
+2021-10-18,1
+2021-10-20,2
+2021-10-22,1
+2021-10-23,2
+2021-10-27,1
+2021-11-01,2
+2021-11-08,2
+2021-11-12,1
+2021-11-13,1
+2021-11-17,1
+2021-11-20,1
+2021-11-25,1
+2021-11-28,1
+2021-12-05,1
+2021-12-15,1
+2021-12-24,2
+2021-12-25,2
+2021-12-26,1
+2021-12-28,1
+2022-01-05,1
+2022-01-07,1
+2022-01-10,1
+2022-01-15,1
+2022-01-28,2
+2022-01-30,1
+2022-02-02,1
+2022-02-13,1
+2022-02-19,1
+2022-02-20,1
+2022-02-22,1
+2022-02-24,1
+2022-03-12,1
+2022-03-19,2
+2022-03-21,1
+2022-03-25,1
+2022-03-26,2
+2022-03-29,1
+2022-04-01,1
+2022-04-03,3
+2022-04-04,3
+2022-04-05,1
+2022-04-06,2
+2022-04-09,1
+2022-04-12,3
+2022-04-14,1
+2022-04-15,1
+2022-04-17,1
+2022-04-18,1
+2022-04-19,1
+2022-04-21,1
+2022-04-22,1
+2022-04-26,1
+2022-05-01,1
+2022-05-03,2
+2022-05-05,2
+2022-05-06,3
+2022-05-07,1
+2022-05-08,1
+2022-05-10,1
+2022-05-11,1
+2022-05-12,1
+2022-05-14,1
+2022-05-15,1
+2022-05-16,1
+2022-05-18,1
+2022-05-19,1
+2022-05-20,2
+2022-05-22,2
+2022-05-23,4
+2022-05-24,1
+2022-05-25,2
+2022-05-26,1
+2022-05-28,1
+2022-05-31,2
+2022-06-01,1
+2022-06-03,1
+2022-06-04,2
+2022-06-05,1
+2022-06-06,2
+2022-06-07,1
+2022-06-08,1
+2022-06-09,1
+2022-06-10,1
+2022-06-11,1
+2022-06-13,1
+2022-06-15,1
+2022-06-17,1
+2022-06-20,3
+2022-06-22,2
+2022-06-24,2
+2022-06-25,2
+2022-06-26,1
+2022-06-27,2
+2022-07-02,2
+2022-07-04,1
+2022-07-05,1
+2022-07-06,1
+2022-07-08,1
+2022-07-09,4
+2022-07-10,2
+2022-07-13,2
+2022-07-14,1
+2022-07-15,1
+2022-07-17,1
+2022-07-18,2
+2022-07-19,2
+2022-07-20,3
+2022-07-21,1
+2022-07-22,2
+2022-07-23,1
+2022-07-24,1
+2022-07-25,1
+2022-07-27,3
+2022-07-29,1
+2022-07-31,1
+2022-08-01,2
+2022-08-02,1
+2022-08-04,1
+2022-08-05,1
+2022-08-06,4
+2022-08-09,1
+2022-08-10,3
+2022-08-12,3
+2022-08-13,1
+2022-08-14,2
+2022-08-16,2
+2022-08-19,2
+2022-08-20,1
+2022-08-21,1
+2022-08-23,2
+2022-08-24,1
+2022-08-25,1
+2022-08-27,1
+2022-08-28,1
+2022-08-29,1
+2022-08-31,1
+2022-09-02,1
+2022-09-03,1
+2022-09-04,1
+2022-09-05,1
+2022-09-09,1
+2022-09-11,1
