+++ advanced-analytics-hybrid/streamlit/data/number_of_users.csv
@@ -1,0 +1,2 @@
+number_of_users
+100
