+++ advanced-analytics-hybrid/streamlit/data/screenviews_by_day.csv
@@ -1,0 +1,151 @@
+date,number_of_screenviews
+2021-09-21,2
+2021-09-25,5
+2021-09-28,6
+2021-09-29,3
+2021-10-02,6
+2021-10-05,5
+2021-10-16,7
+2021-10-17,1
+2021-10-18,4
+2021-10-20,10
+2021-10-22,6
+2021-10-23,11
+2021-10-27,2
+2021-11-01,4
+2021-11-08,10
+2021-11-12,6
+2021-11-13,4
+2021-11-17,7
+2021-11-20,4
+2021-11-25,3
+2021-11-28,2
+2021-12-05,7
+2021-12-15,2
+2021-12-24,7
+2021-12-25,7
+2021-12-26,6
+2021-12-28,2
+2022-01-05,10
+2022-01-07,6
+2022-01-10,1
+2022-01-15,1
+2022-01-28,7
+2022-01-30,7
+2022-02-02,2
+2022-02-13,6
+2022-02-19,9
+2022-02-20,7
+2022-02-22,4
+2022-02-24,4
+2022-03-12,6
+2022-03-19,14
+2022-03-21,3
+2022-03-25,6
+2022-03-26,8
+2022-03-29,2
+2022-04-01,7
+2022-04-03,10
+2022-04-04,5
+2022-04-05,2
+2022-04-06,7
+2022-04-09,2
+2022-04-12,10
+2022-04-14,3
+2022-04-15,8
+2022-04-17,2
+2022-04-18,4
+2022-04-19,6
+2022-04-21,3
+2022-04-22,7
+2022-04-26,2
+2022-05-01,4
+2022-05-03,13
+2022-05-05,12
+2022-05-06,9
+2022-05-07,10
+2022-05-08,1
+2022-05-10,4
+2022-05-11,5
+2022-05-12,2
+2022-05-14,1
+2022-05-15,2
+2022-05-16,6
+2022-05-18,3
+2022-05-19,9
+2022-05-20,10
+2022-05-22,12
+2022-05-23,20
+2022-05-24,6
+2022-05-25,10
+2022-05-26,8
+2022-05-28,5
+2022-05-31,8
+2022-06-01,6
+2022-06-03,2
+2022-06-04,11
+2022-06-05,5
+2022-06-06,7
+2022-06-07,9
+2022-06-08,1
+2022-06-09,3
+2022-06-10,3
+2022-06-11,6
+2022-06-13,6
+2022-06-15,8
+2022-06-17,2
+2022-06-20,17
+2022-06-22,16
+2022-06-24,15
+2022-06-25,7
+2022-06-26,9
+2022-06-27,3
+2022-07-02,6
+2022-07-04,6
+2022-07-05,2
+2022-07-06,1
+2022-07-08,5
+2022-07-09,17
+2022-07-10,4
+2022-07-13,3
+2022-07-14,1
+2022-07-15,1
+2022-07-17,6
+2022-07-18,10
+2022-07-19,14
+2022-07-20,13
+2022-07-21,4
+2022-07-22,16
+2022-07-23,6
+2022-07-24,6
+2022-07-25,1
+2022-07-27,15
+2022-07-29,4
+2022-07-31,7
+2022-08-01,14
+2022-08-02,7
+2022-08-04,9
+2022-08-05,2
+2022-08-06,20
+2022-08-09,5
+2022-08-10,17
+2022-08-12,7
+2022-08-13,3
+2022-08-14,7
+2022-08-16,6
+2022-08-19,12
+2022-08-20,3
+2022-08-21,8
+2022-08-23,15
+2022-08-24,1
+2022-08-25,7
+2022-08-27,9
+2022-08-28,5
+2022-08-29,3
+2022-08-31,2
+2022-09-02,1
+2022-09-03,8
+2022-09-04,3
+2022-09-05,9
+2022-09-09,1
+2022-09-11,5
