+++ advanced-analytics-hybrid/streamlit/data/screenviews_by_screen.csv
@@ -1,0 +1,5 @@
+screen_view_name,number_of_screenviews
+metrics view,2294
+demo view,1618
+additional view,944
+SnowplowSwiftSPMDemo.WebViewController,829
