+++ advanced-analytics-hybrid/streamlit/data/sessions_by_device.csv
@@ -1,0 +1,3 @@
+device_model,number_of_sessions
+"iPhone14,2",110
+"iPhone15,3",100
