+++ advanced-analytics-hybrid/streamlit/data/sessions_by_geo.csv
@@ -1,0 +1,28 @@
+geo_country,number_of_sessions
+US,87
+IN,31
+AU,9
+CA,9
+GB,7
+VN,7
+CH,5
+IL,5
+RO,4
+RU,4
+BR,4
+CN,3
+KR,3
+DE,3
+BD,3
+IR,3
+JP,3
+NO,3
+GR,3
+UA,3
+MV,2
+NZ,2
+PL,2
+SA,2
+SE,1
+IT,1
+DK,1
