+++ advanced-analytics-hybrid/streamlit/queries/sessions/avg_duration.sql
