+++ advanced-analytics-hybrid/streamlit/queries/sessions/avg_duration_by_day.sql
