+++ advanced-analytics-hybrid/streamlit/queries/sessions/num_sessions.sql
