+++ advanced-analytics-hybrid/streamlit/queries/sessions/sessions_by_device.sql
